Add Feeds::ImportWorker worker
